mcp-tools/requirements.txt
# MCP Tools Service Dependencies

# FastMCP (includes Starlette, handles HTTP server internally)
fastmcp>=2.0.0

# HTTP client for Gemini REST API
httpx>=0.26.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0

# Image processing (required for production)
pillow>=10.0.0
pdf2image>=1.16.0

# PDF processing (required for signature cropping)
PyMuPDF>=1.23.0
